requirements.txt
# Dependencias principales del proyecto
pandas
numpy
matplotlib
scikit-learn
Pillow
torch
torchvision
tqdm
PyYAML
optuna